refactor : API 사용 컴포넌트, 파일명 통일
